bugfix + effectBase

+++ Assets/Scripts/Core/Interfaces/IHealable.cs
@@ -2,6 +2,6 @@ namespace Core.Interfaces
 {
     public interface IHealable
     {
-        void Heal(float amount);
+        void Heal(int amount);
     }
 }
